requirements.txt
pandas
wbgapi
psycopg2-binary
matplotlib
seaborn